% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLRtools-package.R
\docType{package}
\name{CLRtools-package}
\alias{CLRtools}
\alias{CLRtools-package}
\title{CLRtools: Diagnostic Tools for Logistic and Conditional Logistic Regression}
\description{
Provides tools for fitting, assessing, and comparing logistic and conditional logistic regression models. Includes residual diagnostics and goodness of fit measures for model development and evaluation in matched case control studies.
}
\author{
\strong{Maintainer}: Brenda Contla Hernández \email{brendacontla11@gmail.com}

Other contributors:
\itemize{
  \item Matthieu Vignes (Provided statistical guidance, conceptual feedback, and supervision) [contributor]
  \item Chris Compton (Provided statistical guidance, conceptual feedback, and supervision) [contributor]
}

}
\keyword{internal}
